[EC-11] feat: Security 기본  설정
