+++ api/src/main/java/org/example/educheck/domain/campusstaff/entity/CampusStaff.java
@@ -1,15 +1,16 @@
-package org.example.educheck.domain.meetingRoom.entity;
+package org.example.educheck.domain.campusstaff.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.example.educheck.domain.campus.Campus;
+import org.example.educheck.domain.member.staff.entity.Staff;
 
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MeetingRoom {
+public class CampusStaff {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -19,5 +20,7 @@ public class MeetingRoom {
     @JoinColumn(name = "campus_id")
     private Campus campus;
 
-    private String name;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "staff_id")
+    private Staff staff;
 }
