+++ api/src/main/java/org/example/educheck/domain/consultingattachmentfile/entity/ConsultingAttachmentAttachmentFile.java
@@ -1,14 +1,15 @@
-package org.example.educheck.domain.consulting.entity;
+package org.example.educheck.domain.consultingattachmentfile.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.educheck.domain.consulting.entity.Consulting;
 
 @Getter
 @Entity(name = "consulting_attachment_file")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AttachmentFile {
+public class ConsultingAttachmentAttachmentFile {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
