+++ api/src/main/java/org/example/educheck/domain/meetingroomreservation/entity/MeetingRoomReservation.java
@@ -1,10 +1,10 @@
-package org.example.educheck.domain.meetingRoomReservation.entity;
+package org.example.educheck.domain.meetingroomreservation.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.educheck.domain.meetingRoom.entity.MeetingRoom;
+import org.example.educheck.domain.meetingroom.entity.MeetingRoom;
 import org.example.educheck.domain.member.entity.Member;
 import org.example.educheck.global.common.entity.BaseTimeEntity;
