+++ api/src/main/resources/application.properties
@@ -3,12 +3,9 @@ spring.datasource.url=jdbc:mysql://${DATABASE_HOST}:${DATABASE_PORT}/${DATABASE_
 spring.datasource.username=${DATABASE_USERNAME}
 spring.datasource.password=${DATABASE_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
 spring.data.redis.host=${REDIS_HOST}
 spring.data.redis.port=${REDIS_PORT}
 spring.data.redis.password=${REDIS_PASSWORD}
-
 spring.docker.compose.lifecycle-management=start-and-stop
-
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
